api: add social accounts to user responses
